Use flexbox instead of absolute position for the CSV export button.
